Move aig to aiger.rs
