+++ sym/src/lib.rs
@@ -1,4 +1,4 @@
-mod aig;
+mod aiger;
 mod buf;
 mod convert;
 mod sym;
